api: KMS envelope key fixes
api::schema::EnvelopeKey is supposed to contain all the data needed to
decrypt the token sensitive contents (i.e. private keys). Since we
currently only support KMS as a decryption/attestation service,
api::schema::EnvelopeKey should be a one-of enum, identifying KMS and
the input data that it requires in order to perform a decryption
operation.

- Moved AWS creds to api::schema::EnvelopeKey::Kms; now the args for
  ApiRequest::AddToken are sane and allow attestation / decryption;
- Changed ApiRequest::RefreshToken to ingest an EnvelopeKey, instead of
  the disparated aws_* creds;
- Added AWS_SESSION_TOKEN to the AWS creds (now stored by EnvelopeKey);
- Added AWS_REGION to the AWS creds (now stored by EnvelopeKey);
- Adapted nitro-evault to the above API changes.

Signed-off-by: Dan Horobeanu <dhr@amazon.com>
Reviewed-by: Alexandru Ciobotaru <alcioa@amazon.com>
Reviewed-by: Bogdan Vasile <bve@amazon.com>
Reviewed-by: Petre Eftime <epetre@amazon.com>