vtok_agent: Update role arn fetch logic to support IAM roles with paths
Currently, the IAM role arn is fetched by using the IMDS metadata. If
the IAM role contains a path, this logic would not work for this case,
as only the role name can be retrieved, not the whole path.

Update the logic to get the IAM role arn by parsing the output of "aws
iam get-role" command. Keep the previous logic for backwards
compatibility, as a new policy has to be set to get the IAM role
metadata using the AWS CLI.

This is an example of how the IAM policy should be set:

{
    "Version": "2012-10-17",
    "Statement": [
        {
            "Effect": "Allow",
            "Action": "iam:GetRole",
            "Resource": "<IAMRoleArn>"
        }
    ]
}

Signed-off-by: Andra Paraschiv <andraprs@amazon.com>