p11ne-cli: improve JSON output reporting
Currently the commands made available by the
p11ne-cli report their output as JSON, by
piping their output through plain `jq`.

This was found to introduce a problem when trying
to process the output in shell scripts (e.g. storing
the output in a shell variable for further processing).
Plain `jq` expects the destination of the JSON
representation to be a terminal.

This commit adds a pretty-print filter to `jq`
(`jq '.'`) such that the output can be processed in
shell scripts as well.

Signed-off-by: Gabriel Bercaru <bercarug@amazon.com>