vtok_p11/db: add x509 certificate support
For each private key provisioned in the token,
associate the optional X509 certificate (or
certificate chain).

Enforce corectness via the the aws-lc crypto backend
and extract several certificate attributes such as:
subject, issuer, serial number and the DER encoded
certificate.

As per general convention the input PEM file has the
certificates concatenated in the order of:
- Server certificate PEM
- Intermediate CAs PEMs
- The root CA PEM

If the certificate chain parsing or verification
fails then the provisioning shall fail.

The certificates are internally labeled in the form
of "acm-ne-cert-{ID}". The CA certificate IDs are
incremental starting from the private key ID while
the token certificate (i.e. associated with the
private key) inherits the key's ID.

Signed-off-by: Alexandru Ciobotaru <alcioa@amazon.com>