crypto/x509: allow x509 chains without root CAs
As per RFC 4158 the chain root-of-trust hierarchy
is based on root CAs or a self-signed certificate.

Some administrators could provision valid chains
but without a self-signed certificate at the end,
causing the verification to fail.

For a provisioned private key the pkcs#11 enclave
makes two assumptions about its certificates:
- if only one certificate, then it is self signed
with this key
- if more certificates, the server certificate is
part of a chain of intermediate CAs

Allow the end-entity (server) certificate to
have its root-of-trust in trusted intermediate
certificates if a self-signed certificate is missing
from the chain.

Signed-off-by: Alexandru Ciobotaru <alcioa@amazon.com>