tests: Create x509 certificate chains on the fly for testing
We have been using a set of fixed certificate chaines for testing for a
long time. Most of the certificates in there are expired or close to
expiring. It has also not been documented what exact cases these chains
have covered.

Replace these fixed certificate chains with creation on the fly when we
run the certificate testing. This has the advantage that we do not run
into the trap of tests failing because certificates run out on some
random date, and at the same time documents the chain setup we are
testing for with these.

I have tried to reconstruct setups that are similar enough to the
existing fixed setup, while trying to keep the number of created certs
minimal and reuse as much as possible.

Signed-off-by: Leonard Foerster <foersleo@amazon.com>