agent: Populate systemd config override when restarting
On AL2 there is a known bug which prevents httpd from working correctly
with PKCS#11 when it is started with -DFOREGROUND option. This option is
set by default in the sample systemd service which gets installed
alongside httpd.

In order to mitigate this problem, override the config explicitly in the
agent code: it was done in this way before until the commit
"c35f99b vtok_agent: Decouple webserver reload from agent" accidentally
removed the httpd systemd configuration code.

So, introduce a function which allows dynamic reconfiguration of the
web-server systemd service when executing post-sync actions. Define the
config for HTTPD (there is already a const string with the required
config in the code).

Since we need to reload the systemd daemon after putting a new config
override, introduce a method to do so.

Clean up unused HTTPD configuration constants.

Signed-off-by: Ivan Orlov <iorlov@amazon.com>