devtool: Bump rustc version
Currently, the project can't be build with rustc 1.63 due to the fact
that some of the dependencies of aws-nitro-enclaves-nsm-api rely on the
newer compiler versions. It generates the following errors:

error: package `once_cell v1.21.3` cannot be built because it requires rustc 1.65 or newer, while the currently activ
e rustc version is 1.63.0
error: package `serde_bytes v0.11.17` cannot be built because it requires rustc 1.68 or newer, while the currently active rustc version is 1.63.0

Bump the default rust toolchain version to 1.71, as specified in the
aws-nitro-enclaves-nsm-api `Config.toml` to be able to build this
subproject. Use the same toolchain version for both parent and enclave
containers. Also, allow re-using RUST_TOOLCHAIN env variable to specify
the toolchain instead of the pre-defined value. This would allow us to
specify the toolchain version as environment variable in CI.

Signed-off-by: Ivan Orlov <iorlov@amazon.com>